requirements.txt
pandas
sqlalchemy
psycopg2-binary
pyarrow
matplotlib
seaborn
python-dotenv